+++ options.go
@@ -44,7 +44,7 @@ func WithStacktrace(enabled bool) Option {
 	}
 }
 
-func WithMaxSize(maxSize int) Option {
+func WithMaxSize(maxSize float64) Option {
 	return func(c *Config) {
 		c.MaxSize = maxSize
 	}
